add rl-dronecan-config
