+++ pyproject.toml
@@ -26,8 +26,9 @@ rl-test-cyphal-specification = "raccoonlab_tools.scripts.cyphal.test_specificati
 rl-test-dronecan-specification = "raccoonlab_tools.scripts.dronecan.test_specification:main"
 rl-check-protocol = "raccoonlab_tools.scripts.common.check_protocol:main"
 rl-upload-firmware = "raccoonlab_tools.scripts.common.upload_firmware:main"
-rl-get-params = "raccoonlab_tools.scripts.dronecan.get_params:main"
-rl-set-params = "raccoonlab_tools.scripts.dronecan.set_params:main"
+rl-get-dronecan-params = "raccoonlab_tools.scripts.dronecan.get_params:main"
+rl-set-dronecan-params = "raccoonlab_tools.scripts.dronecan.set_params:main"
+rl-dronecan-config = "raccoonlab_tools.scripts.dronecan.config:main"
 
 [project.urls]
 "Homepage" = "https://github.com/PonomarevDA/tools"
